build: added individual request and reply examples
